Split requirements into two separate files

+++ backend/initialize_machine.sh
@@ -1,5 +1,6 @@
 #!/usr/bin/env bash
 
+echo "Installing all APT packages..."
 apt install -y make
 if [ ! -e /usr/bin/python3.6 ]
 then
@@ -10,9 +11,12 @@ then
     apt install -y python3-pip
 fi
 
-if [ ! -e /opt/conda/ ]
+enable_hdfs=false  # HDFS is now disabled. Enable it once we will support HDFS!
+if [ ! -e /opt/conda/ && enable_hdfs ]
 then
     # Install Miniconda
+    echo "Installing MiniConda..."
+    echo "WARN: This is a workaround for problems with installing library that handles connection to the HDFS..."
     apt install -y -q curl bzip2
     curl https://repo.continuum.io/miniconda/Miniconda-latest-Linux-x86_64.sh -o /tmp/miniconda.sh
     /bin/bash /tmp/miniconda.sh -b -p /opt/conda
@@ -27,18 +31,23 @@ then
     cp -p /lib/x86_64-linux-gnu/libreadline.so.6 /opt/conda/lib/libreadline.so.6
 fi
 
+echo "Applying environment variables..."
 cd /vagrant/backend
-export PYTHONPATH=`pwd`
+. ./devenv.sh
 
 if [ ! -e /usr/local/bin/docker-compose ]
 then
+    echo "Installing Docker Compose..."
     curl -L https://github.com/docker/compose/releases/download/1.18.0/docker-compose-`uname -s`-`uname -m` -o /usr/local/bin/docker-compose
     chmod +x /usr/local/bin/docker-compose
 fi
+
+echo "Running all Docker dependencies..."
 docker-compose -f /vagrant/docker-compose.yml up -d hbase postgres rabbitmq
 
 echo "Installing all Python packages..."
 make install_packages
+make install_dev_packages
 
 echo "Preparing backend..."
 ./scripts/dev__prepare_backend.sh
